More refactoring of the Slack config to match what we had earlier
